feat: 필터 상단 타이틀 컴포넌트 분리
